架构更新
