--- AI/split_api_code.py
@@ -1,89 +1,0 @@
-import re
-HTML_TAG_PATTERN = re.compile(r'<\s*/?\s*[a-zA-Z][^>]*>')
-
-REGEX_METACHARS = re.compile(r'[*+?^${}()|[\]\\]')  # 正则元字符
-SPLIT_COMMENT_PATTERN = re.compile(r'(?<!:)//')     # 分割行内注释（排除 http://）
-QUOTED_CONTENT_PATTERN = re.compile(r'["\'](.*?)["\']')  # 提取引号内容
-
-# 静态资源黑名单
-BLACK_LIST = (
-    '.png', '.css', '.jpeg', '.jpg', '.gif', '.ico',
-    '.ttf', '.svg', '.woff', '.woff2', '.eot', '.otf',
-    '.mp4', '.mp3', '.m4v', '.aac', '.apk', '.exe',
-)
-
-def has_valid_slash(content: str) -> bool:
-    """判断内容中是否存在至少一个/，其左侧或右侧有数字/字母"""
-    if not isinstance(content, str):
-        return False
-    for i, char in enumerate(content):
-        if char == '/':
-            left_valid = i > 0 and content[i - 1].isalnum()
-            right_valid = i < len(content) - 1 and content[i + 1].isalnum()
-            if left_valid or right_valid:
-                return True
-    return False
-
-
-def extract_relevant_lines(input_str: str) -> str:
-    """
-    从 JS 代码中提取可能包含 API 路径的行（用于送入大模型分析）
-    :param input_str: 原始 JS 代码字符串
-    :return: 提取后的相关行，用换行符连接
-    """
-    if not isinstance(input_str, str):
-        return ""
-
-    relevant_lines = []
-
-    for line in input_str.splitlines():
-        line_stripped = line.strip()
-        # 跳过空行和单行注释
-        if not line_stripped or line_stripped.startswith('//'):
-            continue
-
-        # 跳过单行过长的代码（避免模型处理过长文本）
-        if len(line_stripped) > 800:
-            continue
-
-        # 🚀 快速跳过：不含关键字符的行（提升 3~5 倍性能）
-        if not ('/' in line_stripped or 'http' in line_stripped or 'api' in line_stripped or "=" in line_stripped or ":" in line_stripped):
-            continue
-
-        # 去除行内注释（修复：排除 http:// 中的 //）
-        parts = SPLIT_COMMENT_PATTERN.split(line_stripped, 1)
-        line_no_comment = parts[0].rstrip()
-        if not line_no_comment:
-            continue
-
-        # ===================== 新增逻辑开始 =====================
-
-        # 1️⃣ 去除所有反斜杠
-        line_no_comment = line_no_comment.replace('\\', '')
-
-        # 2️⃣ HTML 标签检测，命中直接跳过
-        if HTML_TAG_PATTERN.search(line_no_comment):
-            continue
-
-        # ===================== 新增逻辑结束 =====================
-
-        # 提取所有引号内的内容
-        quoted_contents = QUOTED_CONTENT_PATTERN.findall(line_no_comment)
-        if not quoted_contents:
-            continue
-
-        # 检查是否有有效路径
-        valid = False
-        for content in quoted_contents:
-            if (has_valid_slash(content) and
-                not REGEX_METACHARS.search(content) and
-                not any(content.lower().endswith(ext) for ext in BLACK_LIST)):
-                valid = True
-                break
-
-        if valid:
-            # 标准化空白字符
-            cleaned_line = ' '.join(line_no_comment.split())
-            relevant_lines.append(cleaned_line)
-
-    return "\n".join(relevant_lines)
