--- ai_security_scanner/utils/__init__.py
@@ -1,5 +1,0 @@
-from .js_formatter import format_code
-from .js_line_extractor import extract_relevant_lines
-
-
-__all__ = ['format_code', 'extract_relevant_lines']
