+++ logger/__init__.py
@@ -1,0 +1,5 @@
+"""工具函数模块"""
+
+from .logger import get_logger
+
+__all__ = ['get_logger']
