+++ processor/analysis/__init__.py
@@ -1,0 +1,5 @@
+from processor.analysis.params.params_scan import AISecurityAuditor
+from processor.js.context.context_extractor import extract_multiple_apis_from_raw_code
+
+
+__all__ = ['AISecurityAuditor', 'extract_multiple_apis_from_raw_code']
