+++ processor/analysis/prompts/__init__.py
@@ -1,0 +1,9 @@
+from .prompts import _SYSTEM_PROMPT_JUDGE, _SYSTEM_PROMPT_ADVISORY
+
+SYSTEM_PROMPT_JUDGE = _SYSTEM_PROMPT_JUDGE
+SYSTEM_PROMPT_ADVISORY = _SYSTEM_PROMPT_ADVISORY
+
+__all__ = [
+    'SYSTEM_PROMPT_JUDGE',
+    'SYSTEM_PROMPT_ADVISORY'
+]
