+++ requirements.txt
@@ -4,50 +4,16 @@ nltk~=3.9.2
 wordninja~=2.0.0
 bs4~=0.0.2
 beautifulsoup4~=4.11.2
-langchain-openai~=0.3.16
 tqdm~=4.67.1
-langchain-community~=0.3.27
-langchain-core~=0.3.74
 jsbeautifier~=1.15.4
 requests~=2.32.5
-rich~=14.1.0
-urllib3~=2.0.7
 user-agent~=0.1.10
-numpy~=1.26.4
-langchain~=0.3.27
 chardet~=5.1.0
 tldextract~=3.4.4
 psutil~=7.0.0
 colorama~=0.4.6
-
-spacy~=3.8.11
 tree-sitter~=0.25.2
-openai~=1.109.1
-cffi~=2.0.0
-defusedxml~=0.7.1
-pillow~=11.0.0
-ipython~=8.12.3
-zlib~=1.2.13
-jupyterlab-pygments~=0.3.0
-pip~=25.1
-attrs~=25.3.0
-distro~=1.9.0
-wheel~=0.45.1
-pysocks~=1.7.1
-protobuf~=6.32.0
-openssl~=3.0.18
-tornado~=6.5.4
-jinja2~=3.1.4
-pyee~=13.0.0
-typing_extensions~=4.15.0
-setuptools~=78.1.1
-markupsafe~=2.1.5
-wcwidth~=0.2.14
-exceptiongroup~=1.3.0
-six~=1.17.0
-python-dateutil~=2.9.0.post0
-greenlet~=3.2.4
-packaging~=25.0
-build~=1.3.0
-platformdirs~=4.5.1
+httpx~=0.28.1
+rich~=14.1.0
 pymysql~=1.1.1
+flask~=3.1.3
